requirements.txt
locust
dotenv
certifi==2019.9.11
chardet==3.0.4
Click==7.0
Flask==1.1.1
gevent==1.4.0
geventhttpclient-wheels==1.3.1.dev2
greenlet==0.4.15
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
locustio==0.12.2
MarkupSafe==1.1.1
msgpack-python==0.5.6
PyYAML==5.1.2
pyzmq==18.1.0
requests==2.22.0
six==1.12.0
urllib3==1.25.6
Werkzeug==0.16.0
python-dotenv
